Bug fixes related to LAST_INSERT_ID()
* 64 bits LAST_INSERT_ID() #1090
* return LAST_INSERT_ID() as type LONGLONG #1092
* set charset as `binary` #1092

+++ include/MySQL_Session.h
@@ -135,7 +135,7 @@ class MySQL_Session
 	char * default_schema;
 
 	uint32_t thread_session_id;
-	unsigned int last_insert_id;
+	unsigned long long last_insert_id;
 	enum session_status status;
 	int healthy;
 	int user_max_connections;
